% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mepsRR}
\alias{mepsRR}
\title{Subset MEPS panel 10 data for RR estimation}
\format{
A \code{data.frame} with columns:
\describe{
\item{DUPERSID}{}
\item{VARSTR}{}
\item{VARPSU}{}
\item{LONGWT}{}
\item{flushot_status}{}
\item{disease_status}{}
}
}
\source{
\href{https://www.meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-106}{MEPS}
}
\usage{
data(mepsRR)
}
\description{
Data from the Medical Expenditure Panel Survey panel 10, modified for the RR study. The data has 9959 observations with 21 variables.
}
\examples{
data(mepsRR)
}
\references{
Assessing the Effectiveness of Influenza Vaccination – influence function approach for comparative studies using publicly available medical expenditure data
(\href{https://www.meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-106}{MEPS})
}
\keyword{datasets}
